Packages: Extract edit post package
