+++ edit-post/plugins.js
@@ -1,0 +1,2 @@
+export * from '../packages/edit-post/src';
+export { default as PluginSidebar } from './components/sidebar/plugin-sidebar';
