+++ packages/edit-post/package.json
@@ -1,0 +1,29 @@
+{
+	"name": "@wordpress/edit-post",
+	"version": "1.0.0-beta.0",
+	"description": "Edit post implementation for WordPress.",
+	"author": "The WordPress Contributors",
+	"license": "GPL-2.0-or-later",
+	"keywords": [
+		"wordpress",
+		"edit-post"
+	],
+	"homepage": "https://github.com/WordPress/gutenberg/tree/master/packages/edit-post/README.md",
+	"repository": {
+		"type": "git",
+		"url": "https://github.com/WordPress/gutenberg.git"
+	},
+	"bugs": {
+		"url": "https://github.com/WordPress/gutenberg/issues"
+	},
+	"main": "build/index.js",
+	"module": "build-module/index.js",
+	"react-native": "src/index",
+	"dependencies": {
+		"@babel/runtime-corejs2": "7.0.0-beta.56",
+		"@wordpress/plugins": "file:../plugins"
+	},
+	"publishConfig": {
+		"access": "public"
+	}
+}
